# Use our own Discard for testing